requirements.txt
openai-whisper>=20231117
assemblyai>=0.33.0
soundfile>=0.12.1